fakedir patch upstreamed
the null buffer issue has been resolved upstreamed with a PR. also, fix
an issue with symlinks/shebangs not being rewritten for the initial
program executed by the startup script -- we need to avoid that call to
execvp happening in the process environment that existed prior to the
injection of the fakedir libraries. to do that, run a new shell where
the exported env vars are set and avoid that shell just execing by first
testing if the fakedir library is present and loaded, then running the
specified executable.